Refactor KasinoGame into separate files
